Handle expressions with pipes (#131)
* test: 💍 Add test for string addition with ngTemplate

When we use string addition with translations, we need to ensure the key
is being extracted correctly. Currently, there were no testcases for
this. This commit adds testcases for the ngTemplate scenario

* feat: 🎸 Allow binary expressions with pipes

When we used binary expressions with pipes, the transloco key was being
ignored by the extract utility. This was because we were not parsing the
Binary expressions from the AST. Add binary expressions with pipes as
transloco supports that.

✅ Closes: #129

* feat: 🎸 Allow method calls with pipes

When we used method calls with pipes, the transloco key was being
ignored by the extract utility. This was because we were not parsing the
MethodCall nodes from the AST. Add method call with pipes as transloco
supports that.

✅ Closes: #129

* feat: 🎸 Allow expressions in ternary condition

When we used ternary conditions with pipes, we were checking for piipes
being used in the true and false cases, but not in the condition side.
Allow the condition to also contain translations as transloco supports
that.